+++ runtime/parachains/src/mock.rs
@@ -293,9 +293,11 @@ impl crate::disputes::SlashingHandler<BlockNumber> for Test {
 }
 
 impl crate::scheduler::Config for Test {
-	type CoreAssigners<T: crate::scheduler::Config> =
-		(crate::scheduler::Parachains, crate::scheduler_parathreads::ParathreadsScheduler);
-	// (crate::scheduler::Parachains, (crate::scheduler_parathreads::ParathreadsScheduler, crate::scheduler_parathreads::ParathreadsScheduler));
+	type CoreAssigners<T: crate::scheduler::Config> = (
+		crate::scheduler_parachains::ParachainsScheduler,
+		crate::scheduler_parathreads::ParathreadsScheduler,
+	);
+	// (crate::scheduler_parachains::Parachains, (crate::scheduler_parathreads::ParathreadsScheduler, crate::scheduler_parathreads::ParathreadsScheduler));
 }
 impl crate::scheduler_parathreads::Config for Test {}
