+++ runtime/polkadot/src/lib.rs
@@ -1322,8 +1322,10 @@ impl parachains_paras_inherent::Config for Runtime {
 impl scheduler_parathreads::Config for Runtime {}
 
 impl parachains_scheduler::Config for Runtime {
-	type CoreAssigners<T: runtime_parachains::scheduler::Config> =
-		(runtime_parachains::scheduler::Parachains, runtime_parachains::scheduler::Parathreads);
+	type CoreAssigners<T: runtime_parachains::scheduler::Config> = (
+		runtime_parachains::scheduler_parachains::ParachainsScheduler,
+		runtime_parachains::scheduler_parathreads::ParathreadsScheduler,
+	);
 }
 
 impl parachains_initializer::Config for Runtime {
